issue-#21: Added long press for menu buttons
Got tired of single button press, especially when changing the tempo, so
now there is long button press support for the menu buttons.

For the "<" and ">" menu button, it keeps executing the prev and next
functions. First with a 500ms delay for the first three iterations, then
it speeds up to 100ms.
For the "Select" menu button, long press will wait 1.5s and for now only
set the LCD to inverse video mode. Later on this will open a separate
configuration menu.

Long press shares the same timer as the chord playback uses. There's
still timer2 available, but that's an 8bit timer, and this kinda needs
16bit timer values - or then an additional counter, but I'll skip that
for now. But this means the button presses will interfere with each
other. I decided the playback has priority here, so the whole long press
mechanism is disabled while chord playback is ongoing.
In that case the menu functions as always, with single button presses.