issue-#19: Added chord list to LCD w/ highlight
Each key's four chord's are now displayed on the LCD for easier
visualization of what's actually available in each key. This should make
it easier to play songs by chord - and also to find out whether a song
is a 4chord song or not.

It felt like a good idea to create a "font" for this instead of using
individual graphics. Using quotes cause it doesn't really make any real
difference - both is just data representing pixels on the LCD.

Added also highlighting the currently played chord. So whenever one of
the four chord buttons is pressed, the corresponding one is lighting up
on the LCD - and goes back to normal when the button is released.