issue-#20: Added metre playback and display support
General support for metre concept added, and included new 3/4 and 6/8
playback next to the regular 4/4 one. Well, at this current point, since
all playback is based on 1/8 notes, the there is no difference between
3/4 and 6/8 apart from LCD content.

Added a font for this as well, could use some improvements still if
something like 12/8 metre should be supported (there's only one digit
supported for now, so.. yeah.)

Also added a little metronome to arpeggio playback (there's no point in
showing a metronome without beat as in the chord playback mode), no idea
if anyone has actual any use to it, but there was this little piece of
LCD space left, and that seemed like a nice to have thing.

Updated the menu, too. Yeah, still using the same old script for the
graphics.c/.h generation..