issue-#25: EEPROM data is validated and sanitized
Ensure the data stored in EEPROM is somewhat sanitized and within the
expected range. If not, the data is overwritten with default fallback
values.

Writing EEPROM data will warn about possible consequences and needs
response before proceeding. Optionally the current EEPROM content is
read back and stored to a file.