Added doc/ directory with local copies of relevant docs and specs
Since the links to specifications were broken, I figured it's maybe best
to just add a local copy just in case. While those links probably won't
ever fully disappear, it's still annoying to end up with a 404 and
having to search for the new location. So now they're also in the repo.